[rtl] Rework accumulate command order of operations
Previously we accumulated by starting with the bias and adding the
values ot it. Now we sum all of the incoming values and then add the
bias. To enable this different adders are used for the operation. It now
reuses the first level of adders that sum the multiply results and the
level below rather than the final level.

This new setup will can also be directly used by the
multiply-accumulate command without having to add further operand
muxing.