[dv] out_compare.py script now reports on concrete matches
Previously is just reported the total matches which includes matches
where we have a don't care in the expected output. Now it gives a
specific count for the matches where there is a concrete value in the
expected output. This is useful for recognising cases where the expected
output has a higher number of don't cares than it should (e.g. the file
is entirely full of don't cares).